Global variable capitalization
Global variables should be capitalized, and use camel capitalization.
Variables are considered global if ever entered into a function via the "global" keyword, or defined outside any function or indenting.
Global variables may be named by identity, with increasing specificity:
ImgProjLazer for the image for the projectile that is a lazer.